search-nlu/requirements.txt
flask==2.2.3
python-dotenv==0.15.0
ibm-watson==7.0.0